+++ core/src/androidJvmMain/kotlin/at/asitplus/awesn1/JcaExtensions.kt
@@ -10,6 +10,7 @@ import at.asitplus.awesn1.encoding.internal.Sink
 import java.math.BigInteger
 import kotlin.experimental.and
 import kotlin.experimental.or
+import kotlin.IgnorableReturnValue
 import kotlin.uuid.ExperimentalUuidApi
 import kotlin.uuid.Uuid
 
@@ -54,6 +55,7 @@ fun Uuid.Companion.fromBigintOrNull(bigInteger: BigInteger): Uuid? =
  */
 @InternalAwesn1Api
 @Throws(IllegalArgumentException::class)
+@IgnorableReturnValue
 fun Sink.writeAsn1VarInt(number: BigInteger): Int {
     if (number == BigInteger.ZERO) { //fast case
         writeByte(0)