+++ core/src/commonMain/kotlin/at/asitplus/awesn1/encoding/internal/NumberEncoding.kt
@@ -10,6 +10,7 @@ import at.asitplus.awesn1.VarUInt.Companion.writeAsn1VarInt
 import at.asitplus.awesn1.encoding.UVARINT_MASK_UBYTE
 import at.asitplus.awesn1.encoding.bitLength
 import at.asitplus.awesn1.encoding.toTwosComplementByteArray
+import kotlin.IgnorableReturnValue
 import kotlin.math.ceil
 
 
@@ -23,6 +24,7 @@ private const val UVARINT_SINGLEBYTE_MAXVALUE_UBYTE: UByte = 0x80u
  */
 
 @InternalAwesn1Api
+@IgnorableReturnValue
 fun Sink.writeAsn1VarInt(number: ULong) = writeAsn1VarInt(number, ULong.SIZE_BITS)
 
 /**
@@ -33,6 +35,7 @@ fun Sink.writeAsn1VarInt(number: ULong) = writeAsn1VarInt(number, ULong.SIZE_BIT
  */
 
 @InternalAwesn1Api
+@IgnorableReturnValue
 fun Sink.writeAsn1VarInt(number: UInt) = writeAsn1VarInt(number.toULong(), UInt.SIZE_BITS)
 
 /**
@@ -43,6 +46,7 @@ fun Sink.writeAsn1VarInt(number: UInt) = writeAsn1VarInt(number.toULong(), UInt.
  */
 
 @InternalAwesn1Api
+@IgnorableReturnValue
 private fun Sink.writeAsn1VarInt(number: ULong, bits: Int): Int {
     if (number == 0uL) { //fast case
         writeByte(0)
@@ -117,6 +121,7 @@ private fun Source<*>.decodeAsn1VarInt(bits: Int): Pair<ULong, ByteArray> {
  * @return the number of byte written to the sink
  */
 @InternalAwesn1Api
+@IgnorableReturnValue
 fun Sink.writeTwosComplementLong(number: Long): Int = appendUnsafe(number.toTwosComplementByteArray())
 
 /**
@@ -124,6 +129,7 @@ fun Sink.writeTwosComplementLong(number: Long): Int = appendUnsafe(number.toTwos
  * @return the number of bytes written
  */
 @InternalAwesn1Api
+@IgnorableReturnValue
 fun Sink.writeTwosComplementULong(number: ULong): Int = appendUnsafe(number.toTwosComplementByteArray())
 
 
@@ -131,6 +137,7 @@ fun Sink.writeTwosComplementULong(number: ULong): Int = appendUnsafe(number.toTw
  * @return the number of bytes written to the sink
  */
 @InternalAwesn1Api
+@IgnorableReturnValue
 fun Sink.writeTwosComplementUInt(number: UInt) = writeTwosComplementLong(number.toLong())
 
 /**
@@ -209,6 +216,7 @@ fun Source<*>.readTwosComplementUInt(nBytes: Int): UInt {
  *  @return the number of bytes written
  */
 @InternalAwesn1Api
+@IgnorableReturnValue
 fun Sink.writeMagnitudeLong(number: Long): Int {
     require(number >= 0)
     return number.toTwosComplementByteArray().let { appendUnsafe(it, if (it[0] == 0.toByte()) 1 else 0) }
@@ -223,6 +231,7 @@ fun Sink.writeMagnitudeLong(number: Long): Int {
  */
 @Throws(IllegalArgumentException::class)
 @InternalAwesn1Api
+@IgnorableReturnValue
 fun Sink.writeAsn1VarInt(number: Asn1Integer): Int {
     require(number is Asn1Integer.Positive) { "Only non-negative numbers are supported" }
     return writeAsn1VarInt(number.uint)